requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
Pillow==9.4.0
protobuf==3.20.3
tk==0.1.0
Werkzeug==2.2.2
